chain_prompt_transmission/src/requirements.txt
# API and Core Dependencies
openai>=1.0.0
python-dotenv>=1.0.0

# Configuration Management
pyyaml>=6.0.0

# Progress Bars and CLI
tqdm>=4.65.0

# Data Validation
jsonschema>=4.17.0

# Testing (Optional but recommended)
pytest>=7.4.0
pytest-cov>=4.1.0

# Type Checking (Optional but recommended)
mypy>=1.5.0
types-PyYAML>=6.0.12
types-tqdm>=4.65.0

# Development Tools (Optional)
black>=23.7.0  # Code formatting
isort>=5.12.0  # Import sorting
flake8>=6.1.0  # Code linting
